+++ ash/src/extensions/ext/tooling_info.rs
@@ -6,17 +6,15 @@ use std::mem;
 
 #[derive(Clone)]
 pub struct ToolingInfo {
-    handle: vk::Instance,
     fns: vk::ExtToolingInfoFn,
 }
 
 impl ToolingInfo {
     pub fn new(entry: &Entry, instance: &Instance) -> Self {
-        let handle = instance.handle();
         let fns = vk::ExtToolingInfoFn::load(|name| unsafe {
-            mem::transmute(entry.get_instance_proc_addr(handle, name.as_ptr()))
+            mem::transmute(entry.get_instance_proc_addr(instance.handle(), name.as_ptr()))
         });
-        Self { handle, fns }
+        Self { fns }
     }
 
     #[doc = "<https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/vkGetPhysicalDeviceToolPropertiesEXT.html>"]
@@ -36,9 +34,5 @@ impl ToolingInfo {
 
     pub fn fp(&self) -> &vk::ExtToolingInfoFn {
         &self.fns
-    }
-
-    pub fn instance(&self) -> vk::Instance {
-        self.handle
     }
 }