+++ ash/src/extensions/khr/dynamic_rendering.rs
@@ -5,17 +5,15 @@ use std::mem;
 
 #[derive(Clone)]
 pub struct DynamicRendering {
-    handle: vk::Device,
     fns: vk::KhrDynamicRenderingFn,
 }
 
 impl DynamicRendering {
     pub fn new(instance: &Instance, device: &Device) -> Self {
-        let handle = device.handle();
         let fns = vk::KhrDynamicRenderingFn::load(|name| unsafe {
-            mem::transmute(instance.get_device_proc_addr(handle, name.as_ptr()))
+            mem::transmute(instance.get_device_proc_addr(device.handle(), name.as_ptr()))
         });
-        Self { handle, fns }
+        Self { fns }
     }
 
     #[doc = "<https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/vkCmdBeginRenderingKHR.html>"]
@@ -39,9 +37,5 @@ impl DynamicRendering {
 
     pub fn fp(&self) -> &vk::KhrDynamicRenderingFn {
         &self.fns
-    }
-
-    pub fn device(&self) -> vk::Device {
-        self.handle
     }
 }