+++ ash/src/extensions/khr/get_physical_device_properties2.rs
@@ -7,17 +7,15 @@ use std::ptr;
 
 #[derive(Clone)]
 pub struct GetPhysicalDeviceProperties2 {
-    handle: vk::Instance,
     fns: vk::KhrGetPhysicalDeviceProperties2Fn,
 }
 
 impl GetPhysicalDeviceProperties2 {
     pub fn new(entry: &Entry, instance: &Instance) -> Self {
-        let handle = instance.handle();
         let fns = vk::KhrGetPhysicalDeviceProperties2Fn::load(|name| unsafe {
-            mem::transmute(entry.get_instance_proc_addr(handle, name.as_ptr()))
+            mem::transmute(entry.get_instance_proc_addr(instance.handle(), name.as_ptr()))
         });
-        Self { handle, fns }
+        Self { fns }
     }
 
     #[doc = "<https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/vkGetPhysicalDeviceFeatures2KHR.html>"]
@@ -154,9 +152,5 @@ impl GetPhysicalDeviceProperties2 {
 
     pub fn fp(&self) -> &vk::KhrGetPhysicalDeviceProperties2Fn {
         &self.fns
-    }
-
-    pub fn instance(&self) -> vk::Instance {
-        self.handle
     }
 }