+++ ash/src/extensions/khr/push_descriptor.rs
@@ -6,17 +6,15 @@ use std::mem;
 
 #[derive(Clone)]
 pub struct PushDescriptor {
-    handle: vk::Device,
     fns: vk::KhrPushDescriptorFn,
 }
 
 impl PushDescriptor {
     pub fn new(instance: &Instance, device: &Device) -> Self {
-        let handle = device.handle();
         let fns = vk::KhrPushDescriptorFn::load(|name| unsafe {
-            mem::transmute(instance.get_device_proc_addr(handle, name.as_ptr()))
+            mem::transmute(instance.get_device_proc_addr(device.handle(), name.as_ptr()))
         });
-        Self { handle, fns }
+        Self { fns }
     }
 
     #[doc = "<https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/vkCmdPushDescriptorSetKHR.html>"]
@@ -62,9 +60,5 @@ impl PushDescriptor {
 
     pub fn fp(&self) -> &vk::KhrPushDescriptorFn {
         &self.fns
-    }
-
-    pub fn device(&self) -> vk::Device {
-        self.handle
     }
 }